portfolio provision uses pydantic classes that validate the fields defined for csp data and csp result classes
