+++ Pipfile
@@ -31,6 +31,7 @@ msrestazure = "*"
 azure-mgmt-authorization = "*"
 transitions = "*"
 sqlalchemy-json = "*"
+pydantic = "*"
 
 [dev-packages]
 bandit = "*"
