+++ atst/models/mixins/state_machines.py
@@ -1,4 +1,5 @@
 from enum import Enum
+import importlib
 
 from atst.database import db
 
@@ -96,3 +97,25 @@ class FSMMixin():
     def prepare_reset(self, event): pass
     def before_reset(self, event): pass
     def after_reset(self, event): pass
+
+    def _stage_to_classname(self, stage):
+        return "".join(map(lambda word: word.capitalize(), stage.replace('_', ' ').split(" ")))
+
+    def stage_csp_class(self, stage, class_type):
+        """
+        given a stage name and class_type return the class
+        class_type is either 'payload' or 'result'
+
+        """
+        cls_name = "".join([self._stage_to_classname(stage), "CSP", class_type.capitalize()])
+        try:
+            return getattr(importlib.import_module("atst.domain.csp.cloud"), cls_name)
+        except AttributeError:
+            print("could not import CSP Result class <%s>" % cls_name)
+
+    def fail_stage(self, stage):
+        getattr(self.machine, 'fail_'+stage)()
+
+    def finish_stage(self, stage):
+        self.trigger('finish_'+stage)
+
